+++ ante/expected_keeper.go
@@ -4,13 +4,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// TreasuryKeeper for tax charging & recording
-type TreasuryKeeper interface {
-	RecordEpochTaxProceeds(ctx sdk.Context, delta sdk.Coins)
-	GetTaxRate(ctx sdk.Context) (taxRate sdk.Dec)
-	GetTaxCap(ctx sdk.Context, denom string) (taxCap sdk.Int)
-}
-
 // OracleKeeper for feeder validation
 type OracleKeeper interface {
 	ValidateFeeder(ctx sdk.Context, feederAddr sdk.AccAddress, validatorAddr sdk.ValAddress) error
